be/pyproject.toml
[tool.poetry]
name = "be"
version = "0.1.0"
description = ""
authors = ["vvsch <vvschrus@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.3"
django-rest-framework = "^0.1.0"
django-filter = "^23.2"
drf-dynamic-fields = "^0.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
